botao add e remove da pagina detalhes
