+++ pokedex/src/Components/Cards/CardPokemons.js
@@ -36,9 +36,6 @@ const CardPokemons = () => {
     if (index === -1) {
       const novaPokedex = [...states.pokedex, pokemon];
       setters.setPokedex(novaPokedex);
-      //logica para sumir com o card depois que ele é adicionado na pokedex
-      document.getElementById(pokemon.name).style.display = 'none';
-      alert("to na pokedex!!!")
     }
   };
 
@@ -58,6 +55,7 @@ const CardPokemons = () => {
                   <BotaoAdicionar
                     adicionarPokemon={adicionarPokemon}
                     pokemon={pokemon}
+                    nome={"Adicionar"}
                   />
                   <BotaoDetalhes url={pokemon.name} />
                 </div>
